n80xx-common: update include and inherit-product
if we have two files with the same name used with PRODUCT_COPY_FILES,
it takes always the first one and not the ones in the tree

This allows us to override files from smdk4412-common

  * use include instead -include wherever appropriate
    . when using "-include", make ignores a makefile which does not
      exist, with no error message
    . use "include" for all makefiles which are essential for a
      working build

  * remove include for vendor board product
    . that file not exists

The ueventd file in n80xx is smaller and maybe older than the one in smdk4412-common.
I think it's better to keep using the same file, so I've copied it to n80xx-common
without further changes (I've only removed fmradio dev path) to avoid potential
issues for now and allow future device specific changes.

Signed-off-by: Chirayu Desai<cdesai@cyanogenmod.org>
Signed-off-by: Humberto Borba <humberos@gmail.com>
Signed-off-by: marco9333 <marco.magliona@gmail.com>
Change-Id: I27ead596ad14a9f16e2c18b532b2894f3f26218d